feat: 병합오류수정, MapSharedViewModel분리
