+++ app/src/main/java/nbc/group/recipes/presentation/fragment/BottomSheetFragment.kt
@@ -17,20 +17,20 @@ import kotlinx.coroutines.launch
 import nbc.group.recipes.presentation.BottomSheetAdapter
 import nbc.group.recipes.databinding.FragmentBottomsheetBinding
 import nbc.group.recipes.presentation.MainActivity
-import nbc.group.recipes.viewmodel.MapViewModel
+import nbc.group.recipes.viewmodel.MapSharedViewModel
 
 @AndroidEntryPoint
 class BottomSheetFragment : BottomSheetDialogFragment() {
     private val binding get() = _binding!!
     private var _binding: FragmentBottomsheetBinding? = null
 
-    private val sharedMapViewModel : MapViewModel by activityViewModels()
+    private val mapSharedViewModel : MapSharedViewModel by activityViewModels()
 
     private val bottomSheetAdapter : BottomSheetAdapter by lazy {
         BottomSheetAdapter{ item, position ->
             // 클릭시 실행할 동작
             // sharedMapViewModel을 통해서 클릭한 위치의 특산물 데이터를 전달!!
-            sharedMapViewModel.getSelectedSpecialty(item)
+            mapSharedViewModel.getSelectedSpecialty(item)
             Log.d("click_specialtyData__",item.toString())
 
             // RecipeFragment로 이동
@@ -68,7 +68,7 @@ class BottomSheetFragment : BottomSheetDialogFragment() {
     private fun observeSpecialtyData(){
         viewLifecycleOwner.lifecycleScope.launch{
 
-            sharedMapViewModel.specialtie.flowWithLifecycle(viewLifecycleOwner.lifecycle, Lifecycle.State.STARTED).collectLatest{
+            mapSharedViewModel.specialtie.flowWithLifecycle(viewLifecycleOwner.lifecycle, Lifecycle.State.STARTED).collectLatest{
                 Log.d("it_region_data__", it.toString())    // "검색한 지역의 특산물"값 받아옴
 
                 // 어뎁터 업데이트
