+++ app/src/main/java/nbc/group/recipes/viewmodel/MainViewModel.kt
@@ -103,7 +103,7 @@ class MainViewModel @Inject constructor(
     fun getItem(string: String) {
         viewModelScope.launch {
             _specialties.emit(
-                repository.getSpecialty("") // 전체 지역
+                repository.getAreaName("") // 전체 지역
             )
         }
     }
