+++ app/src/main/java/nbc/group/recipes/viewmodel/MapViewModel.kt
@@ -16,7 +16,7 @@ import javax.inject.Inject
 @HiltViewModel
 class MapViewModel @Inject constructor(
     private val searchRepository: SearchRepository,
-    private val repository: RecipeSpecialtyRepository,
+
 ) : ViewModel() {
 
     // 검색어
@@ -25,25 +25,6 @@ class MapViewModel @Inject constructor(
 
     fun getRegionSearch(query: String) = viewModelScope.launch {
         _regionSearch.emit(searchRepository.requestSearch(query))
-    }
-
-
-
-    // 지역이름
-    private val _specialtie = MutableStateFlow<SpecialtyResponse?>(null)
-    val specialtie : StateFlow<SpecialtyResponse?> = _specialtie
-
-    fun getSpecialtie(ariaName : String) = viewModelScope.launch {
-            _specialtie.emit(repository.getSpecialty(ariaName))
-        }
-
-
-    // 클릭한 특산물 데이터
-    private val _selectedSpecialty = MutableStateFlow<Item?>(null)
-    val selectedSpecialty: StateFlow<Item?> = _selectedSpecialty
-
-    fun getSelectedSpecialty(item: Item?) {
-        _selectedSpecialty.value = item
     }
 
 }
